Added Id column to categories management, Fixed Delete Issue

+++ resources/js/views/Categories.vue
@@ -1,73 +1,73 @@
 <template>
-    <v-card>
-      <v-card-title>
-        Categories
-        <v-spacer></v-spacer>
-        <v-text-field
-          v-model="search"
-          append-icon="mdi-magnify"
-          label="Search"
-          single-line
-          hide-details
-        ></v-text-field>
-      </v-card-title>
-      <v-data-table
-        :headers="headers"
-        :items="categories"
-        :search="search"
-        :loading="loading"
-        loading-text="Loading Categories... Please wait"
-      >
-        <template v-slot:item.action="{ item }">
-          <v-icon small class="mr-2" @click="editItem(item)">mdi-pencil</v-icon>
-          <v-icon small @click="deleteItem(item)">mdi-delete</v-icon>
-        </template>
-        <template v-slot:no-data>
-          <v-btn color="primary" @click="initialize">Reset</v-btn>
-        </template>
-      </v-data-table>
-      <v-dialog v-model="dialog" max-width="500px">
-        <template v-slot:activator="{ on }">
-          <v-btn color="primary" dark class="mb-2" v-on="on">New Item</v-btn>
-        </template>
-        <v-card>
-          <v-card-title>
-            <span class="headline">{{ formTitle }}</span>
-          </v-card-title>
+  <v-card>
+    <v-card-title>
+      Categories
+      <v-spacer></v-spacer>
+      <v-text-field
+        v-model="search"
+        append-icon="mdi-magnify"
+        label="Search"
+        single-line
+        hide-details
+      ></v-text-field>
+    </v-card-title>
+    <v-data-table
+      :headers="headers"
+      :items="categories"
+      :search="search"
+      :loading="loading"
+      loading-text="Loading Categories... Please wait"
+    >
+      <template v-slot:item.action="{ item }">
+        <v-icon small class="mr-2" @click="editItem(item)">mdi-pencil</v-icon>
+        <v-icon small @click="deleteItem(item)">mdi-delete</v-icon>
+      </template>
+      <template v-slot:no-data>
+        <v-btn color="primary" @click="initialize">Reset</v-btn>
+      </template>
+    </v-data-table>
+    <v-dialog v-model="dialog" max-width="500px">
+      <template v-slot:activator="{ on }">
+        <v-btn color="primary" dark class="mb-2" v-on="on">New Item</v-btn>
+      </template>
+      <v-card>
+        <v-card-title>
+          <span class="headline">{{ formTitle }}</span>
+        </v-card-title>
 
-          <v-card-text>
-            <v-container>
-              <v-row>
-                <v-col cols="12" md="6">
-                  <v-text-field v-model="editedItem.name" label="Category Name"></v-text-field>
-                </v-col>
-                <v-col cols="12" md="6">
-                  <v-select
-                    v-model="editedItem.measurement"
-                    :items="categories"
-                    label="Category"
-                    item-text="name"
-                    item-value="id"
-                    return-object
-                    prepend-icon="mdi-view-dashboard"
-                  ></v-select>
-                </v-col>
-              </v-row>
-            </v-container>
-          </v-card-text>
+        <v-card-text>
+          <v-container>
+            <v-row>
+              <v-col cols="12" md="6">
+                <v-text-field v-model="editedItem.name" label="Category Name"></v-text-field>
+              </v-col>
+              <v-col cols="12" md="6">
+                <v-select
+                  v-model="editedItem.measurement"
+                  :items="categories"
+                  label="Category"
+                  item-text="name"
+                  item-value="id"
+                  return-object
+                  prepend-icon="mdi-view-dashboard"
+                ></v-select>
+              </v-col>
+            </v-row>
+          </v-container>
+        </v-card-text>
 
-          <v-card-actions>
-            <v-spacer></v-spacer>
-            <v-btn color="blue darken-1" text @click="close">Cancel</v-btn>
-            <v-btn color="blue darken-1" text @click="save">Save</v-btn>
-          </v-card-actions>
-        </v-card>
-      </v-dialog>
-      <v-snackbar v-model="snackbar" :timeout="snackbarTimeout">
-        {{ snackbarText }}
-        <v-btn color="blue" text @click="snackbar = false">Close</v-btn>
-      </v-snackbar>
-    </v-card>
+        <v-card-actions>
+          <v-spacer></v-spacer>
+          <v-btn color="blue darken-1" text @click="close">Cancel</v-btn>
+          <v-btn color="blue darken-1" text @click="save">Save</v-btn>
+        </v-card-actions>
+      </v-card>
+    </v-dialog>
+    <v-snackbar v-model="snackbar" :timeout="snackbarTimeout">
+      {{ snackbarText }}
+      <v-btn color="blue" text @click="snackbar = false">Close</v-btn>
+    </v-snackbar>
+  </v-card>
 </template>
 
 <script>
@@ -77,11 +77,12 @@ export default {
       search: "",
       headers: [
         {
-          text: "Name",
+          text: "id",
           align: "left",
           sortable: false,
-          value: "name"
+          value: "id"
         },
+        { text: "Name", value: "name" },
         { text: "Created At", value: "created_at" },
         { text: "Updated At", value: "updated_at" },
         { text: "Actions", value: "action", sortable: false }
@@ -137,11 +138,13 @@ export default {
 
     deleteItem(item) {
       const index = this.categories.indexOf(item);
-      confirm("Are you sure you want to delete this category?") &&
-        this.categories.splice(index, 1);
-      axios.delete("api/categories/" + item.id);
-      this.snackbarText = "Category deleted";
-      this.snackbar = true;
+      if (confirm("Are you sure you want to delete this category?")) {
+        if (this.categories.splice(index, 1)) {
+          axios.delete("api/categories/" + item.id);
+          this.snackbarText = "Category deleted";
+          this.snackbar = true;
+        }
+      }
     },
 
     close() {
@@ -159,8 +162,7 @@ export default {
         this.snackbar = true;
         axios
           .patch("api/categories/" + this.editedItem.id, {
-            name: this.editedItem.name,
-            measurement_id: this.editedItem.measurement.id
+            name: this.editedItem.name
           })
           .then(function(response) {
             console.log(response);
@@ -168,15 +170,14 @@ export default {
       } else {
         axios
           .post("api/categories", {
-            name: this.editedItem.name,
-            measurement_id: this.editedItem.measurement.id
+            name: this.editedItem.name
           })
           .then(function(response) {
             console.log(response);
           });
         this.categories.push({
           name: this.editedItem.name,
-          "measurement.name": this.editedItem.measurement.name
+          "category.name": this.editedItem.category.name
         });
         this.snackbar = true;
         this.snackbarText = "Category created";
